Move classes to Iteration namespace
