+++ tests/Iteration/SampleIterator.php
@@ -1,9 +1,9 @@
 <?php
 declare( strict_types = 1 );
 
-namespace PHP\Tests\ForEachLoops;
+namespace PHP\Tests\Iteration;
 
-use PHP\ForEachLoops\Iterator;
+use PHP\Iteration\Iterator;
 
 class SampleIterator extends Iterator
 {
